close #23 
close #22 
close #20 
close #24 
close #12 

done

+++ flask_todo/routes.py
@@ -1,26 +1,46 @@
-from flask import render_template, url_for, redirect, request
+from flask import url_for, redirect, jsonify, request
 from flask_todo import app, db
-from flask_todo.form import CreateTodoForm, LoginForm, SignUpForm
-from flask_todo.models import Todo
+from flask_todo.models import Todo, TodoSchema
+import json
+import datetime
+from pytz import timezone
 
 
-@app.route('/', methods=['GET', 'POST'])
-@app.route('/create', methods=['GET', 'POST'])
+@app.route('/index', methods=('GET',))
 def index():
-    form = CreateTodoForm()
-    if form.validate_on_submit():
-        input_form_data = request.form['todo_name']
-        todo = Todo(title=input_form_data)
-        db.session.add(todo)
-        db.session.commit()
-        return redirect(url_for('index'))
-
     todos = db.session.query(Todo).all()
-    return render_template('index.html', form=form, todos=todos)
+    for todo in todos:
+        dt_naive_to_utc_replace = todo.date_posted.replace(
+            tzinfo=datetime.timezone.utc)
+        todo.date_posted = dt_naive_to_utc_replace.astimezone(
+            timezone('Asia/Tokyo'))
+
+    return jsonify({'status': 'ok',
+                    'todos': TodoSchema(many=True).dump(todos)})
 
 
-@app.route('/<int:id>delete', methods=('POST',))
-def delete(id):
+@app.route('/create', methods=('POST',))
+def create():
+    if not request.is_json:
+        return jsonify({"message": "Missing JSON in request"}), 400
+
+    data = request.data.decode('utf-8')
+    data = json.loads(data)
+    title = str(data['title'])
+    todo = Todo(title=title)
+    db.session.add(todo)
+    db.session.commit()
+    return redirect(url_for('index'))
+
+
+@app.route('/delete', methods=('POST',))
+def delete():
+    if not request.is_json:
+        return jsonify({"message": "Missing JSON in request"}), 400
+
+    data = request.data.decode('utf-8')
+    data = json.loads(data)
+    id = str(data['delete_id'])
     db.session.query(Todo).filter(Todo.id == id).delete()
     db.session.commit()
     return redirect(url_for('index'))
@@ -28,15 +48,9 @@ def delete(id):
 
 @app.route('/login', methods=['GET', 'POST'])
 def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        pass
-    return render_template('login.html', form=form)
+    return 'OK'
 
 
 @app.route('/signup', methods=['GET', 'POST'])
 def signup():
-    form = SignUpForm()
-    if form.validate_on_submit():
-        pass
-    return render_template('signup.html', form=form)
+    return 'OK'
